requirements.txt
python-dotenv
pandas 
pyodbc
pydantic
pymilvus
langchain-google-genai
google-generativeai
google-ai-generativelanguage
ipywidgets
ipykernel
streamlit
sentence_transformers
numpy
fastapi
uvicorn
chromadb